+++ deployment/aws-filebased-config/cdk/constructs/internal/grants/kms_test.go
@@ -12,7 +12,6 @@ import (
 )
 
 func TestGrantKMSEfsCmkUse(t *testing.T) {
-	defer jsii.Close()
 	stack, role := newTestStack(t)
 	key := awskms.Key_FromKeyArn(stack, jsii.String("K"),
 		jsii.String("arn:aws:kms:us-east-1:111122223333:key/abcd1234-12ab-34cd-56ef-1234567890ab"))